feat Deletion: can't delete yourself as admin anymore

+++ app/Actions/Jetstream/DeleteUser.php
@@ -12,6 +12,10 @@ class DeleteUser implements DeletesUsers
      */
     public function delete(User $user): void
     {
+        if ($user->isRole('admin') &&
+            count(User::all()->where('role', '=', 'admin')) <= 1
+        ) return;
+
         $user->deleteProfilePhoto();
         $user->tokens->each->delete();
         $user->delete();
